@@ -15,16 +15,16 @@
   },
   "devDependencies": {
     "@anatine/zod-openapi": "^1.12.1",
-    "@sveltejs/kit": "~1.15.10",
+    "@sveltejs/kit": "~1.16.3",
     "@types/node": "~18.16.3",
     "@types/uuid": "^9.0.1",
-    "@typescript-eslint/eslint-plugin": "~5.59.2",
-    "@typescript-eslint/parser": "~5.59.2",
+    "@typescript-eslint/eslint-plugin": "~5.59.5",
+    "@typescript-eslint/parser": "~5.59.5",
     "carbon-components": "^10.58.3",
     "carbon-components-svelte": "~0.73.5",
     "carbon-icons-svelte": "^11.4.0",
     "carbon-preprocess-svelte": "^0.9.1",
-    "eslint": "~8.39.0",
+    "eslint": "~8.40.0",
     "eslint-config-prettier": "~8.8.0",
     "eslint-plugin-svelte3": "~4.0.0",
     "http-status-codes": "^2.2.0",
@@ -34,14 +34,14 @@
     "prettier-plugin-svelte": "~2.10.0",
     "rxjs": "^7.8.1",
     "sass": "~1.62.1",
-    "svelte": "~3.58.0",
+    "svelte": "~3.59.1",
     "svelte-adapter-firebase": "~0.15.0",
-    "svelte-check": "~3.2.0",
+    "svelte-check": "~3.3.2",
     "sveltekit-flash-message": "~0.11.3",
     "tslib": "~2.5.0",
-    "type-fest": "~3.9.0",
+    "type-fest": "~3.10.0",
     "typescript": "~5.0.4",
-    "vite": "~4.3.4",
+    "vite": "~4.3.5",
     "zod": "~3.21.4",
     "zod-form-data": "~2.0.1"
   },
