@@ -1,12 +1,40 @@
-<script>
+<script lang="ts">
   import { ProgressIndicator, ProgressStep } from 'carbon-components-svelte';
 
-  import { page } from '$app/stores';
-  import { goto } from '$app/navigation';
   import { browser } from '$app/environment';
+  import { goto } from '$app/navigation';
+  import { page } from '$app/stores';
   import Seo from '$lib/components/Seo.svelte';
+  import { derived } from 'svelte/store';
+  import { z } from 'zod';
 
-  const indexArray = [
+  const ParamsSchema = z.object({
+    platformId: z.string().optional(),
+    gameId: z.string().optional(),
+    platformGameMiniGameId: z.string().optional(),
+    platformGameMiniGameModeId: z.string().optional(),
+    platformGameMiniGameModeStageId: z.string().optional(),
+  });
+  const DataSchema = z.object({
+    platform: z.object({ name: z.string() }).optional(),
+    game: z.object({ name: z.string() }).optional(),
+    platformGameMiniGame: z.object({ miniGameName: z.string() }).optional(),
+    platformGameMiniGameMode: z.object({ modeName: z.string() }).optional(),
+    platformGameMiniGameModeStage: z.object({ stageName: z.string() }).optional(),
+  });
+
+  type ParamKey = keyof z.infer<typeof ParamsSchema>;
+  interface Index {
+    key: ParamKey;
+    index: number;
+  }
+
+  const state = derived(page, ({ data, params }) => ({
+    data: DataSchema.parse(data),
+    params: ParamsSchema.parse(params),
+  }));
+
+  const indexArray: readonly Index[] = [
     { key: 'platformId', index: 0 },
     { key: 'gameId', index: 1 },
     { key: 'platformGameMiniGameId', index: 2 },
@@ -14,7 +42,7 @@
     { key: 'platformGameMiniGameModeStageId', index: 4 },
   ];
 
-  $: currentIndex = indexArray.find(({ key }) => !$page.params[key])?.index ?? 5;
+  $: currentIndex = indexArray.find(({ key }) => !$state.params[key])?.index ?? 5;
 </script>
 
 <h1>Submit a new score</h1>
@@ -26,14 +54,24 @@
   preventChangeOnClick={!browser}
 >
   {@const {
-    platformId,
-    gameId,
-    platformGameMiniGameId,
-    platformGameMiniGameModeId,
-    platformGameMiniGameModeStageId,
-  } = $page.params}
+    params: {
+      platformId,
+      gameId,
+      platformGameMiniGameId,
+      platformGameMiniGameModeId,
+      platformGameMiniGameModeStageId,
+    },
+    data: {
+      platform,
+      game,
+      platformGameMiniGame,
+      platformGameMiniGameMode,
+      platformGameMiniGameModeStage,
+    },
+  } = $state}
   <ProgressStep
     label="Platform"
+    secondaryLabel={platform?.name}
     complete={!!platformId}
     on:click={() => {
       goto('/b/score/submit');
@@ -41,6 +79,7 @@
   />
   <ProgressStep
     label="Game"
+    secondaryLabel={game?.name}
     complete={!!gameId}
     disabled={!platformId}
     on:click={() => {
@@ -49,6 +88,7 @@
   />
   <ProgressStep
     label="Mini game"
+    secondaryLabel={platformGameMiniGame?.miniGameName}
     complete={!!platformGameMiniGameId}
     disabled={!gameId}
     on:click={() => {
@@ -57,6 +97,7 @@
   />
   <ProgressStep
     label="Mode"
+    secondaryLabel={platformGameMiniGameMode?.modeName}
     complete={!!platformGameMiniGameModeId}
     disabled={!platformGameMiniGameId}
     on:click={() => {
@@ -65,6 +106,7 @@
   />
   <ProgressStep
     label="Stage"
+    secondaryLabel={platformGameMiniGameModeStage?.stageName}
     complete={!!platformGameMiniGameModeStageId}
     disabled={!platformGameMiniGameModeId}
     on:click={() => {
@@ -73,7 +115,7 @@
       );
     }}
   />
-  <ProgressStep label="Score" disabled={!$page.params.platformGameMiniGameModeStageId} />
+  <ProgressStep label="Score" disabled={!platformGameMiniGameModeStageId} />
 </ProgressIndicator>
 
 <Seo title="Submit a new score" description="Submit a new score" />
@@ -89,5 +131,10 @@
 
   main {
     margin-top: 1rem;
+  }
+
+  main :global(h2) {
+    margin-top: 2rem;
+    margin-bottom: 2rem;
   }
 </style>
