@@ -1,26 +1,19 @@
 import {
   GAME_MINI_GAME_END_POINT,
-  PLATFORM_END_POINT,
   PLATFORM_GAME_MINI_GAME_END_POINT,
 } from '$env/static/private';
 import { arrayUniqBy } from '$lib/array-uniq-by';
 import { httpServer } from '$lib/server/http-server';
+import { PlatformService } from '$lib/server/services/platform.service';
 import { error } from '@sveltejs/kit';
 import { z } from 'zod';
 import type { PageServerLoad } from './$types';
 
 export const load = (async ({ fetch }) => {
+  const platformService = PlatformService.create(fetch);
   const [platformsResponse, gameMiniGamesResponse, platformGameMiniGamesResponse] =
     await Promise.all([
-      httpServer(PLATFORM_END_POINT, {
-        fetch,
-        schema: z.array(
-          z.object({
-            platformId: z.string(),
-            name: z.string(),
-          })
-        ),
-      }),
+      platformService.get(),
       httpServer(GAME_MINI_GAME_END_POINT, {
         fetch,
         schema: z.array(
