@@ -1,10 +1,10 @@
 import { parseFormData } from '$lib/server/form-data';
+import { PlatformService } from '$lib/server/services/platform.service';
+import { fail, redirect } from '@sveltejs/kit';
+import { StatusCodes } from 'http-status-codes';
+import { z } from 'zod';
 import { zfd } from 'zod-form-data';
 import type { Actions } from './$types';
-import { z } from 'zod';
-import { fail, redirect } from '@sveltejs/kit';
-import { httpServer } from '$lib/server/http-server';
-import { PLATFORM_END_POINT } from '$env/static/private';
 
 export const actions = {
   default: async ({ fetch, request }) => {
@@ -18,19 +18,11 @@ export const actions = {
     if (formDataError) {
       return fail(formDataError.status, { error: formDataError });
     }
-    const [responseError] = await httpServer(PLATFORM_END_POINT, {
-      fetch,
-      schema: z.object({
-        platformId: z.string(),
-        name: z.string(),
-        shortName: z.string(),
-      }),
-      method: 'POST',
-      body: formData,
-    });
+    const platformService = PlatformService.create(fetch);
+    const [responseError] = await platformService.post(formData);
     if (responseError) {
       return fail(responseError.status, { error: responseError });
     }
-    throw redirect(301, '/a/platform');
+    throw redirect(StatusCodes.MOVED_PERMANENTLY, '/a/platform');
   },
 } satisfies Actions;
