@@ -1,15 +1,7 @@
 <script lang="ts">
   import { browser } from '$app/environment';
   import { enhance } from '$app/forms';
-  import {
-    Button,
-    RadioTile,
-    StructuredList,
-    StructuredListBody,
-    StructuredListCell,
-    StructuredListRow,
-    TileGroup,
-  } from 'carbon-components-svelte';
+  import { Button, RadioTile, TileGroup } from 'carbon-components-svelte';
   import { ChevronLeft, ChevronRight } from 'carbon-icons-svelte';
 
   export let data;
@@ -31,22 +23,7 @@
   }
 </script>
 
-<h2>Parameters selected</h2>
-
-<StructuredList condensed class="no-margin-bottom">
-  <StructuredListBody>
-    <StructuredListRow>
-      <StructuredListCell>Platform</StructuredListCell>
-      <StructuredListCell>{data.platform.name}</StructuredListCell>
-    </StructuredListRow>
-    <StructuredListRow>
-      <StructuredListCell>Game</StructuredListCell>
-      <StructuredListCell>{data.game.name}</StructuredListCell>
-    </StructuredListRow>
-  </StructuredListBody>
-</StructuredList>
-
-<h3>Select the mini game</h3>
+<h2>Select the mini game</h2>
 
 <form method="POST" use:enhance>
   <TileGroup bind:selected={platformGameMiniGameIdSelected}>
@@ -77,14 +54,3 @@
     Next
   </Button>
 </form>
-
-<style lang="scss">
-  h2 {
-    margin-bottom: 1rem;
-  }
-
-  h3 {
-    margin-bottom: 2rem;
-    margin-top: 1rem;
-  }
-</style>
